+++ doc/release-notes.md
@@ -1,4 +1,5 @@
 #1.1.3
+- patient saymyname readded for calleeservice call       
 
 #1.1.2
 - upgrade to Spring Boot 2.7.0